feat: quoteId (#19)
At Kaala Wallet's request, we added a quoteId field to the quote endpoint response types, and for the transfer endpoint request types.

This means a wallet should be able to show an end-user a fiat-to-crypto exchange rate that the provider has committed to when they initiate a transfer. Without it, rates and fees would need to be estimated until the transfer clears.

For a provider, this means that an identifier for a quote must be issued alongside the quote. When a wallet tries to create a request with this quoteId, if the quote has not yet expired, the provider must honor the exchange rate and fees from the original quote.

NOTE: guaranteedUntil is now a required field for Quote responses (otherwise offering quoteId's would be of little use)