docs(fee): disambiguate fees (#78)
Previously, the meaning of fee was ambiguous because cryptoAmount and fiatAmount denoted an exchange rate that was singular for the quote, but fee could vary per account type if multiple account types are accepted for the same quote.

This PR disambiguates the meaning of fee by making it singular as well (one fee per quote).

As a side effect, providers will not be able to offer different fees per account type for the same quote. However, this was already the case because of how cryptoAmount and fiatAmount are defined, as the exact amount in/out for a transfer (fees and all!). This spec change is designed to avoid a breaking change on the CICO provider side by simply moving an optional field to a different place within the same quote response.

An alternative approach would be to move cryptoAmount and fiatAmount to where fee previously was, nested under fiatAccount in the quote response. This would allow for different fees per fiat account type, but would create a breaking change that is onerous for wallets and, likely, CICO providers. I'd prefer not to do this, since I'm not aware of any providers who rely on being able to levy different fees for different account types right now anyway. Of course, this is always an option for a v2 version of the spec.