fix(webhooks)!: use v1= for hmac signatures (#83)
BREAKING CHANGE: use v1= instead of s= for webhook signatures

This matches the Persona webhook signing scheme better, which is linked
to in that section. We (Valora) and at least one other group (Bitssa)
have already accidentally implemented webhook request signing and
validation this way, probably because the linked-to example uses it.