feat(user actions): Add URLUserAction schema (#99)
This PR adds in a new `URLUserAction` User Action Details Schema, which
has the same schema as the existing `PSEUserAction` schema, but lacks
the semantics of being explicitly for the PSE payment system. As a
result, the `PSEUserAction` schema is being deprecated; it is still
supported by the FiatConnect spec, but its use is discouraged.