fix(user action details): Swagger fixes for User Action Details (#100)
This PR contains a correction to the Swagger spec around User Action
Details, regarding the UserActionDetails object passed to the
/transfer/in endpoint, which was missing certain fields, causing false
negatives when running validations. (Note that the solution here is sort
of a hack; IIRC Swagger 2.0 doesn't allow unions of other types, so I
needed to make an ad-hoc union and make all the fields optional, which
isn't ideal.)

This also fixes the ToC links in the specification itself.