feat(user-action-details): add AccountNumberUserAction (#102)
We've seen FiatConnect providers use the URLUserActionDetails schema and
direct users to their own website just to explain to them that they need
to send funds to the same CICO provider's bank account. This defies the
intention of the UserActionDetails feature, which is supposed to give
FiatConnect clients a chance to explain to users what steps are needed
to complete a transfer in-- not just hand them off to complete arbitrary
steps on some other website. This PR adds a UserActionDetails schema
that should bridge the gap, allowing CICO providers to solicit a bank
transfer from end-users without redirecting them to another website.